refactor: adopt clean architecture pattern and standardize JPA persistence mappings (#1)
* feat: refactor architecture and standardize service-layer pattern

- Removed repository usage from all controllers
- Introduced service layer across all modules
- Enforced Controller → Service → Repository architecture
- Refactored AdminBound module with dependency checks
- Migrated entity creation to Lombok @Builder
- Fixed constructor/type mismatch issues across modules
- Standardized DTO-based request handling
- Centralized entity resolution in service layer
- Refactored WorkOrder module with clean lifecycle handling
- Delegated transition logic to state machine service
- Removed transition log creation from WorkOrderService
- Refactored Feeder module with transactional assignment logging
- Refactored Survey module to use service-driven flow
- Moved validation logic (GPS + sequence) into service layer
- Standardized API responses using ApiResponse wrapper
- Cleaned and simplified all controllers

* refactor: migrate to BIGINT IDs and enforce clean architecture

- Standardized all entity IDs to Long (BIGINT alignment)
- Refactored controllers to use service layer only
- Introduced service layer across modules
- Implemented WorkOrder state machine with DB-driven transitions
- Added transition logging and validation rules
- Replaced constructors with Lombok builders
- Cleaned imports and removed wildcard usage

* feat(backend): add warehouse module, fix WO assignment flow, map exceptions, add tests

Map all custom exceptions in GlobalExceptionHandler (400/403/409)
Refactor WO assignment to use /assign + state machine (CREATED → ASSIGNED)
Add warehouse module (categories, item types APIs)
Add core validator & state machine tests
Update API contract and README

* fix: use Hibernate @CreationTimestamp/@UpdateTimestamp to prevent NOT NULL violations on created_at/updated_at/performed_at
explicitly name persistence unit "hesco"

* refactor: align service/controller to DTO pattern and fix persistence mappings

- moved entity-to-DTO mapping into service layer
- simplified controllers to delegate only
- resolved LazyInitializationException by mapping within transaction
- fixed column naming mismatches (snake_case mapping)
- added proper PostgreSQL enum handling using Hibernate 6 support

* Commented the exception printstacktrace statement.

---------

+++ backend/hesco/src/main/java/com/lmkr/hesco/workorder/entity/WorkOrderStatus.java
@@ -22,7 +22,7 @@ public class WorkOrderStatus {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short id;
+    private Long id;
 
     @Column(nullable = false, unique = true, length = 40)
     private String code;
